had to rework the code a bit
As I continued working on #define, I eventually ran into the ever-annoying declaration-related errors.
So I broke the "directives" file into a header and implementation (which I figured would happen anyway
as the project grew, plus it's just the right way to do things).  But knowing I had to deal with nested
#includes, I moved a lot of the code out of main.  Which broke include(), which meant I had to refactor
a bit.  Anyway, include is working as before, so now I can move on. :)